conversion_nb_scrpt/modularized/requirements.txt
pyspark==3.4.1
boto3==1.28.52
psycopg2-binary==2.9.6
python-dotenv==1.0.0